I add functionality in AdminPrice

+++ src/components/ui/AdminComponents/ColorsAdmin/ColorsAdmin.tsx
@@ -5,26 +5,37 @@ import { deleteColor } from '../../../../cruds/crudColor'
 import { useStoreModal } from '../../../../store/useStoreModal'
 import { AdminColor } from '../../Modals/AdminColor/AdminColor'
 import { useStoreColor } from '../../../../store/useStoreColor'
+import { IProduct } from '../../../../types/IProduct'
+import { getAllProducts } from '../../../../cruds/crudProduct'
 
 export const ColorsAdmin = () => {
 
     const {modalAdminColor, openModalAdminColor} = useStoreModal()
+    const [products, setProducts] = useState<IProduct[]>()
     
-    const [selectedColor, setSelectedColor] = useState<IColor>()
-    const {colors, fetchColors} = useStoreColor()
+    const {colors, fetchColors, setActivecolor} = useStoreColor()
 
     useEffect(() => {
-        
+        const getProducts = async() => {
+            const productsFetched = await getAllProducts()
+            setProducts(productsFetched)
+        }
+        getProducts()
         fetchColors()
     },[])
 
 
-    const openModalColorEdit = (color : IColor) => {
-        openModalAdminColor(2)
-        setSelectedColor(color)
-    }
-
     const handleDelete = async(colorId : string) => {
+
+        const colorInProduct = products?.some(products => 
+            products.colors.some(color => color.id === Number(colorId))
+        )
+
+        if(colorInProduct){
+            alert('El color se encuentra asignado a un producto')
+            return
+        }
+
         try {
             await deleteColor(colorId)
             alert('Se elimino el color')
@@ -34,6 +45,11 @@ export const ColorsAdmin = () => {
             console.log(error.message);
             
         }
+    }
+
+    const handleEdit = (color : IColor) => {
+        openModalAdminColor(2)
+        setActivecolor(color)
     }
 
     return (
@@ -70,7 +86,7 @@ export const ColorsAdmin = () => {
                         
                         <td>
                             <div className={styles.actionButtons}>
-                                <button onClick={() => openModalColorEdit(color)}>Editar</button>
+                                <button onClick={() => handleEdit(color)}>Editar</button>
                                 <button onClick={() => handleDelete(String(color.id))}>Eliminar</button>
                             </div>
                         </td>
@@ -79,7 +95,7 @@ export const ColorsAdmin = () => {
                 </tbody>
             </table>
         </div>
-        {modalAdminColor.type && <div className={styles.modalBackdrop}> <AdminColor color={selectedColor}/></div>}
+        {modalAdminColor.type && <div className={styles.modalBackdrop}> <AdminColor/></div>}
         </div>
     )
 }
